+++ Mathematics/Math.cs
@@ -7,9 +7,9 @@ namespace AdventOfCode_2019.Mathematics
             return (value < 0) ? value * -1 : value;
         }
 
-        public static int[] ToDigitArray(int value)
+        public static long[] ToDigitArray(long value)
         {
-            int[] result = new int[(int)System.Math.Log10(value) + 1];
+            long[] result = new long[(long)System.Math.Log10(value) + 1];
             for (int i = result.Length - 1; i >= 0; i--) 
             {
                 result[i] = value % 10;
